+++ crates/diagnostics/src/types.rs
@@ -32,4 +32,5 @@ pub enum BibError {
     ExpectingEq,
     ExpectingFieldValue,
     UnusedEntry,
+    DuplicateEntry(Vec<TextRange>),
 }